refactor codebase - resolves #6
