+++ src/interfaces/editorConfigProvider.ts
@@ -1,0 +1,7 @@
+import * as editorconfig from 'editorconfig';
+import {TextDocument} from 'vscode';
+
+export interface IEditorConfigProvider {
+	getSettingsForDocument(document: TextDocument): editorconfig.knownProps;
+	getDefaultSettings(): any;
+}
