+++ src/interfaces/editorSettings.ts
@@ -1,0 +1,5 @@
+// Can be removed once https://github.com/Microsoft/vscode/issues/2797 is resolved
+export interface EditorSettings {
+	tabSize: string | number,
+	insertSpaces: string | boolean
+}
